Split behavior impl collection tests
