+++ src/typechecker/tests/resolver_collection/behavior_impls/impl_metadata.rs
@@ -1,0 +1,215 @@
+use super::*;
+
+#[test]
+fn collect_declarations_with_symbols_uses_resolver_behavior_impl_metadata() {
+    let mut program = parse_program(
+        r#"
+Json<T>: behavior {
+    encode: (Self) T
+}
+
+Point: { x: i32 }
+
+Point.implements(Json<str>) {
+    encode = (value: Point) str { "point" }
+}
+"#,
+    );
+    let symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    if let Declaration::ImplBlock {
+        behavior_type_args, ..
+    } = &mut program.declarations[2]
+    {
+        behavior_type_args[0] = AstType::I32;
+    }
+    let mut tc = TypeChecker::new();
+
+    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
+
+    assert!(
+        tc.behavior_impls
+            .contains(&("Point".to_string(), "Json_str".to_string())),
+        "resolver metadata should restore the validated Json<str> impl"
+    );
+    assert!(
+        !tc.behavior_impls
+            .contains(&("Point".to_string(), "Json_i32".to_string())),
+        "AST-only Json<i32> impl drift should not remain after resolver collection"
+    );
+    assert!(
+        tc.diagnostics.is_empty(),
+        "resolver-restored impl metadata should avoid stale AST impl diagnostics: {:?}",
+        tc.diagnostics
+    );
+}
+
+#[test]
+fn collect_declarations_with_symbols_does_not_fallback_to_stale_ast_behavior_impl_metadata() {
+    let mut program = parse_program(
+        r#"
+Json<T>: behavior {
+    encode: (Self) T
+}
+
+Point: { x: i32 }
+
+Point.implements(Json<str>) {
+    encode = (value: Point) str { "point" }
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_behavior_impl_refs_for_test(Namespace::Type, "Point", None);
+    if let Declaration::ImplBlock {
+        behavior_type_args, ..
+    } = &mut program.declarations[2]
+    {
+        behavior_type_args[0] = AstType::I32;
+    }
+    let mut tc = TypeChecker::new();
+
+    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
+
+    assert!(
+            !tc.behavior_impls
+                .contains(&("Point".to_string(), "Json_i32".to_string())),
+            "resolver-backed collection should not keep AST-only behavior impl refs when resolver impl metadata is incomplete"
+        );
+}
+
+#[test]
+fn collect_declarations_with_symbols_does_not_synthesize_stale_impl_defaults_after_target_restore()
+{
+    let mut program = parse_program(
+        r#"
+Point: { x: i32 }
+Json: behavior {
+    encode: (self: Self) str { "default" }
+}
+
+Point.implements(Json) {
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_behavior_impl_refs_for_test(Namespace::Type, "Point", None);
+    if let Declaration::ImplBlock {
+        type_name,
+        behavior,
+        ..
+    } = &mut program.declarations[2]
+    {
+        *type_name = "Missing".to_string();
+        *behavior = Some("AlsoMissing".to_string());
+    }
+    let mut tc = TypeChecker::new();
+
+    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
+
+    assert!(
+            !tc.behavior_impls
+                .contains(&("Point".to_string(), "Json".to_string())),
+            "resolver-backed collection should not keep AST-only behavior impl refs when resolver impl metadata is incomplete"
+        );
+    assert!(
+        !tc.methods.contains_key("Missing.encode"),
+        "resolver-backed default synthesis should not keep stale AST target method keys"
+    );
+    assert!(
+            !tc.methods.contains_key("Point.encode"),
+            "resolver-backed default synthesis should not synthesize behavior defaults when resolver impl metadata is incomplete"
+        );
+    assert!(
+            tc.diagnostics.is_empty(),
+            "resolver-backed collection should not validate stale AST-only impl refs after target restoration when resolver impl metadata is incomplete: {:?}",
+            tc.diagnostics
+        );
+}
+
+#[test]
+fn collect_declarations_with_symbols_uses_resolver_behavior_impl_name_metadata() {
+    let mut program = parse_program(
+        r#"
+Json<T>: behavior {
+    encode: (Self) T
+}
+
+Point: { x: i32 }
+
+Point.implements(Json<str>) {
+    encode = (value: Point) str { "point" }
+}
+"#,
+    );
+    let symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    if let Declaration::ImplBlock { behavior, .. } = &mut program.declarations[2] {
+        *behavior = Some("Missing".to_string());
+    }
+    let mut tc = TypeChecker::new();
+
+    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
+
+    assert!(
+        tc.diagnostics.is_empty(),
+        "resolver-restored impl name metadata should avoid stale AST impl diagnostics: {:?}",
+        tc.diagnostics
+    );
+}
+
+#[test]
+fn collect_declarations_with_symbols_uses_resolver_behavior_impl_target_and_name_metadata() {
+    let mut program = parse_program(
+        r#"
+Json<T>: behavior {
+    encode: (Self) T
+}
+
+Point: { x: i32 }
+
+Point.implements(Json<str>) {
+    encode = (value: Point) str { "point" }
+}
+"#,
+    );
+    let symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    if let Declaration::ImplBlock {
+        type_name,
+        behavior,
+        behavior_type_args,
+        ..
+    } = &mut program.declarations[2]
+    {
+        *type_name = "Missing".to_string();
+        *behavior = Some("AlsoMissing".to_string());
+        behavior_type_args[0] = AstType::I32;
+    }
+    let mut tc = TypeChecker::new();
+
+    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
+
+    assert!(
+        tc.behavior_impls
+            .contains(&("Point".to_string(), "Json_str".to_string())),
+        "resolver metadata should restore the validated Point implements Json<str> association"
+    );
+    assert!(
+            !tc.behavior_impls
+                .contains(&("Missing".to_string(), "AlsoMissing_i32".to_string())),
+            "stale AST-only impl target and behavior metadata should not remain after resolver collection"
+        );
+    assert!(
+            tc.diagnostics.is_empty(),
+            "resolver-restored impl target and behavior metadata should avoid stale AST impl diagnostics: {:?}",
+            tc.diagnostics
+        );
+}
